requirements.txt
requests
pandas
python-dotenv
websocket-client
pymongo
schedule
kafka-python
